Remkonoteboom/4.3

+++ src/tactic/ui/container/tab_wdg.py
@@ -27,7 +27,17 @@ class TabWdg(BaseRefreshWdg):
             'description': 'show the + button',
              'values': 'true|false',
             'category': 'Display'
-        }
+        },
+        'show_context_menu': {
+            'description': 'show the context menu',
+             'values': 'true|false',
+            'category': 'Display'
+        },
+        'show_remove': {
+            'description': 'show the close button',
+             'values': 'true|false',
+            'category': 'Display'
+        },
     }
 
     def get_onload_js(my):
@@ -107,6 +117,30 @@ spt.tab.get_header = function(name) {
             return headers[i];
         }
     }
+    return null;
+}
+
+
+
+spt.tab.get_content = function(name) {
+    var top = spt.tab.top;
+    var tab_id = top.getAttribute("spt_tab_id");
+
+    var content_top = top.getElement(".spt_tab_content_top");
+    var all_contents = content_top.getElements(".spt_tab_content");
+
+    // FIXME: this breaks when opening new tabs for some reason
+    //return all_contents;
+
+    for (var i = 0; i < all_contents.length; i++ ) {
+        var content_tab_id = all_contents[i].getAttribute("spt_tab_id");
+        var content_name = all_contents[i].getAttribute("spt_element_name");
+        if (content_name == name) {
+            return all_contents[i];
+        }
+    }
+
+
     return null;
 }
 
@@ -773,6 +807,64 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
 
 }
 
+spt.tab.close = function(src_el) {
+    // src_el should be a child of spt_tab_content or spt_tab_header
+    if (!src_el) {
+        spt.error('src_el passed in to spt.tab.close() does not exist.');
+        return;
+    }
+    spt.tab.top = src_el.getParent(".spt_tab_top");
+    var top = spt.tab.top;
+    var headers = spt.tab.get_headers();
+    if (headers.length == 1) {
+        return;
+    }
+
+
+    var content = src_el.getParent(".spt_tab_content");
+    var header;
+    var element_name;
+    // check if it's a header child
+    if (content) {
+        element_name = content.getAttribute("spt_element_name");
+        header = spt.tab.get_selected_header(element_name);
+    } else {
+
+        header = src_el.getParent(".spt_tab_header");
+        if (header) {
+            element_name = header.getAttribute("spt_element_name");
+            content = spt.tab.get_content(element_name);
+        }
+
+    }
+    if (!header || !content) {
+        spt.error('Tab close cannot find the header or content. Abort');
+        return;
+    }
+
+    
+    var opener = header.getAttribute("spt_tab_opener");
+    var element_name = header.getAttribute("spt_element_name");
+    header.destroy();
+
+    content.destroy();
+
+    var last_element_name = spt.tab.get_last_selected_element_name();
+    last_element_name = null;
+
+    // make the opener active
+    if (opener) {
+        spt.tab.select(opener);
+    }
+    else if (last_element_name) {
+        spt.tab.select(last_element_name);
+    }
+    else {
+        var last = headers[headers.length - 1].getAttribute("spt_element_name");
+        spt.tab.select(last);
+    }
+}
+
         '''
 
 
@@ -878,7 +970,6 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
         inner = DivWdg();
         top.add(inner);
 
-
         if not Container.get_dict("JSLibraries", "spt_tab"):
             inner.add_behavior( {
             'type': 'load',
@@ -897,6 +988,7 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
         #outer_header.add(header_div)
         header_div.add_style("height: 30px")
         header_div.add_class("spt_tab_header_top")
+        header_div.add_style("overflow-y: hidden")
         #header_div.add_style("width: 5000")
         header_div.add_style("float: left")
 
@@ -1025,7 +1117,6 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
 
         content_top = DivWdg()
 
-
         # add a div so that it breaks correctly
         if my.mode == 'default':
             content_top.add("<div style='height:5px'></div>")
@@ -1033,11 +1124,46 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
             palette = content_top.get_palette()
             border = palette.color("border")
             content_top.add_style("border: 1px solid %s" % border)
-            content_top.add_style("margin-top: -5px")
 
         inner.add(content_top)
         content_top.add_class("spt_tab_content_top")
         content_top.add_style("min-height: 500px")
+
+        height = my.kwargs.get("height")
+        #height = 600
+        #height = None
+        if height:
+            content_top.add_style("height: %s" % height)
+            content_top.add_style("overflow-y: auto")
+            #content_top.add_style("overflow-x: hidden")
+        """
+        else:
+            content_top.add_style("overflow-y: auto")
+            content_top.add_style("border: solid 1px red")
+            content_top.add_behavior( {
+                'type': 'load',
+                'unique_id': my.unique_id,
+                'cbjs_action': '''
+                var el = $(bvr.unique_id);
+                if (!el) {
+                    return;
+                }
+                //el.setStyle("border", "solid 1px blue");
+                var size = el.getSize();
+                bvr.src_el.setStyle("height", size.y);
+                bvr.src_el.setStyle("max-height", size.y);
+                bvr.src_el.setStyle("border", "solid 1px blue");
+
+                window.onresize = function() {
+                    var size = el.getSize();
+                    console.log(size);
+                    bvr.src_el.setStyle("height", size.y);
+                    bvr.src_el.setStyle("max-height", size.y);
+                }
+
+                '''
+            } )
+        """
 
         width = my.kwargs.get("width")
         if not width:
@@ -1046,8 +1172,29 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
             content_top.add_style("min-width: %s" % width)
 
         content_top.add_class("tab_content_top")
-        content_top.add_color("color", "color")
-        content_top.add_color("background", "background")
+
+        color_mode = my.kwargs.get("color_mode")
+        if color_mode == "transparent":
+            pass
+        else:
+            content_top.add_color("color", "color")
+            content_top.add_color("background", "background")
+
+
+        """
+        content_top.add_behavior( {
+            'type': 'load',
+            'cbjs_action': '''
+            new Scrollable(bvr.src_el);
+            '''
+        } )
+        content_top.add_style("overflow: hidden")
+        content_top.add_style("height: 300px")
+        content_top.add_style("padding-right: 15px" )
+        """
+
+
+
 
         # put in a content box for each element
         for element_name in element_names:
@@ -1158,16 +1305,16 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
 
         icon_div = DivWdg()
         icon_div.add_style("padding: 0 2px 0 2px")
-        icon_div.set_round_corners(3, corners=['TR','TL'])
+        icon_div.set_round_corners(12, corners=['TR'])
         from tactic.ui.widget import IconButtonWdg
         icon = IconButtonWdg(title="New Tab", icon=IconWdg.PLUS)
         icon = IconWdg("New Tab", IconWdg.PLUS)
         #icon.add_style("top: -1px")
         #icon.add_style("left: 0px")
         #icon.add_style("position: absolute")
-        icon.add_style("margin-top: -1px")
         icon.add_style("margin-left: 3px")
         icon_div.add_class("hand")
+        icon_div.add_style("opacity: 0.5")
 
         icon_div.add(icon)
         icon.add_behavior( {
@@ -1177,15 +1324,17 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
         spt.tab.add_new();
         '''
         } )
-        icon_div.add_style("padding-top: 4px")
 
         icon_div.add_style("float: left")
-        icon_div.add_style("height: 20px")
-        icon_div.add_style("width: 18px")
-        icon_div.add_style("margin-left: 2px")
-        icon_div.add_gradient("background", "background")
-        #icon_div.add_gradient("background", "tab_background", default="background")
-        icon_div.add_border()
+        icon_div.add_style("margin-top: 2px")
+        icon_div.add_style("padding-top: 4px")
+        icon_div.add_style("height: 21px")
+        icon_div.add_style("width: 22px")
+        icon_div.add_style("margin-left: 4px")
+        icon_div.add_gradient("background", "background", -5, 5)
+        icon_div.add_style("border-style: solid")
+        icon_div.add_style("border-width: 1px 1px 0px 1px")
+        icon_div.add_color("border-color", "border")
         icon_div.add_style("text-align: center")
         div.add(icon_div);
 
@@ -1208,7 +1357,7 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
             icon_div.add_style("height: 20px")
             icon_div.add_style("width: 10px")
             icon_div.add_style("margin-left: -1px")
-            icon_div.add_gradient("background", "background")
+            icon_div.add_gradient("background", "background", -5, 5)
             icon_div.add_border()
             icon_div.add_style("text-align: center")
             div.add(icon_div);
@@ -1248,7 +1397,7 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
         icon_div.add_style("height: 20px")
         icon_div.add_style("width: 18px")
         icon_div.add_style("margin-left: 2px")
-        icon_div.add_gradient("background", "background")
+        icon_div.add_gradient("background", "background", -5, 5)
         icon_div.add_border()
         icon_div.add_style("text-align: center")
         div.add(icon_div);
@@ -1327,8 +1476,13 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
             var activator = spt.smenu.get_activator(bvr);
             var top = activator.getParent(".spt_tab_top");
 
+            // add new if this is the last oni
+            var headers = spt.tab.get_headers();
+            if (headers.length == 1) {
+                spt.tab.add_new();
+            }
+            
             spt.tab.top = top;
-            spt.tab.add_new();
 
             var header = activator;
             var element_name = header.getAttribute("spt_element_name");
@@ -1341,7 +1495,7 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
                     spt.panel.load_popup_with_html( element_name, content.innerHTML );
                     spt.behavior.destroy_element(content);
                 }
-            }
+            } 
 
             '''
         } )
@@ -1365,8 +1519,11 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
 
             var class_name = header.getAttribute("spt_class_name");
             var kwargs_str = header.getAttribute("spt_kwargs");
-            var kwargs = JSON.parse(kwargs_str);
-
+            var kwargs = {};
+            if (kwargs_str) {
+                kwargs_str = kwargs_str.replace(/&quote;/g, '"');
+                kwargs = JSON.parse(kwargs_str);
+            }
             var contents = spt.tab.get_contents();
             for (var i=0; i<contents.length; i++) {
                 var content = contents[i];
@@ -1384,31 +1541,31 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
 
 
 
-        menu_item = MenuItem(type='separator')
-        menu.add(menu_item)
 
+        if my.kwargs.get("show_remove") not in ['false', False]: 
+            menu_item = MenuItem(type='separator')
+            menu.add(menu_item)
+            menu_item = MenuItem(type='action', label='Close Tab')
+            menu_item.add_behavior( {
+                'cbjs_action': '''
+                var activator = spt.smenu.get_activator(bvr);
+                var top = activator.getParent(".spt_tab_top");
+                spt.tab.top = top;
 
-        menu_item = MenuItem(type='action', label='Close Tab')
-        menu_item.add_behavior( {
-            'cbjs_action': '''
-            var activator = spt.smenu.get_activator(bvr);
-            var top = activator.getParent(".spt_tab_top");
-            spt.tab.top = top;
+                var header = activator;
+                var element_name = header.getAttribute("spt_element_name");
+                spt.behavior.destroy_element(header);
 
-            var header = activator;
-            var element_name = header.getAttribute("spt_element_name");
-            spt.behavior.destroy_element(header);
-
-            var contents = top.getElements(".spt_tab_content");
-            for (var i=0; i<contents.length; i++) {
-                var content = contents[i];
-                if (content.getAttribute("element_name") == element_name) {
-                    spt.behavior.destroy_element(content);
+                var contents = top.getElements(".spt_tab_content");
+                for (var i=0; i<contents.length; i++) {
+                    var content = contents[i];
+                    if (content.getAttribute("element_name") == element_name) {
+                        spt.behavior.destroy_element(content);
+                    }
                 }
-            }
-            '''
-        } )
-        menu.add(menu_item)
+                '''
+            } )
+            menu.add(menu_item)
 
 
 
@@ -1446,20 +1603,21 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
                 */
          
 
-
+                var br = '\n';
                 var xml = '';
-                xml += '<element>\n';
-                xml += '  <display class="'+class_name+'">\n';
+                xml += '<element>' + br;
+                xml += '  <display class="'+class_name+'">'  + br;
                 for (var name in kwargs) {
                   if (name == 'class_name') {
                     continue;
                   }
-                  xml += '    <'+name+'>'+kwargs[name]+'</'+name+'>\n';
+                  xml += '    <'+name+'>'+kwargs[name]+'</'+name+'>' + br;
                 }
-                xml += '  </display>\n';
-                xml += '</element>\n';
+                xml += '  </display>' + br;
+                xml += '</element>';
 
-                spt.alert(xml);
+                var html = spt.convert_to_html_display(xml);
+                spt.alert(html, {type:'html'});
                 '''
             } )
             menu.add(menu_item)
@@ -1543,22 +1701,15 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
         header.add_style("border-style: solid")
         header.add_style("border-color: %s" % border)
         header.add_style("border-width: 1px 1px 0px 1px")
+        header.add_style("overflow: hidden")
 
         header.add_style("float: left")
-        header.add_style("padding: 5px")
+        header.add_style("padding: 7px 5px")
         header.add_style("margin-right: 1px")
         #header.add_style("margin-left: 1px")
         if is_IE:
             header.add_style("width: 150px")
         header.add_class("hand")
-
-        #line = DivWdg()
-        #header.add(line)
-        #line.add_style("height: 1px")
-        #line.add_style("width: 100%")
-        #line.add_style("background: red")
-        #line.add("&nbsp;")
-        #line.add_style("margin-top: -5px")
 
         if is_selected:
             header.add_color("color", "color")
@@ -1645,11 +1796,13 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
 
         remove_wdg.add_styles("float: right; position: relative; padding-right: 14px")
         from pyasm.widget import IconButtonWdg
-        icon = IconButtonWdg("Remove Tab", IconWdg.CLOSE_INACTIVE)
+        #icon = IconButtonWdg("Remove Tab", IconWdg.CLOSE_INACTIVE)
+        icon = IconWdg("Remove Tab", "BS_REMOVE", opacity=0.3)
         icon.add_class("spt_icon_inactive")
         icon.add_styles("margin: auto;position: absolute;top: 0;bottom: 0; max-height: 100%")
         remove_wdg.add(icon)
-        icon = IconButtonWdg("Remove Tab", IconWdg.CLOSE_ACTIVE)
+        #icon = IconButtonWdg("Remove Tab", IconWdg.CLOSE_ACTIVE)
+        icon = IconWdg("Remove Tab", "BS_REMOVE")
         icon.add_class("spt_icon_active")
         icon.add_style("display: none")
         icon.add_styles("margin: auto;position: absolute;top: 0;bottom: 0; max-height: 100%")
@@ -1675,43 +1828,7 @@ spt.tab.header_drag_action = function( evt, bvr, mouse_411) {
         remove_wdg.add_behavior( {
         'type': 'click_up',
         'cbjs_action': '''
-            spt.tab.top = bvr.src_el.getParent(".spt_tab_top");
-            var top = spt.tab.top;
-            var headers = spt.tab.get_headers();
-            if (headers.length == 1) {
-                return;
-            }
-
-            var header = bvr.src_el.getParent(".spt_tab_header");
-            var opener = header.getAttribute("spt_tab_opener");
-            var element_name = header.getAttribute("spt_element_name");
-            header.destroy();
-
-            var content_top = top.getElement(".spt_tab_content_top");
-            var contents = content_top.getElements(".spt_tab_content");
-            for (var i = 0; i < contents.length; i++ ) {
-                var content = contents[i];
-                if (content.getAttribute("spt_element_name") == element_name) {
-                    content.destroy();
-                    break;
-                }
-            }
-
-            var last_element_name = spt.tab.get_last_selected_element_name();
-            last_element_name = null;
-
-            // make the opener active
-            if (opener) {
-                spt.tab.select(opener);
-            }
-            else if (last_element_name) {
-                spt.tab.select(last_element_name);
-            }
-            else {
-                var headers = spt.tab.get_headers();
-                var last = headers[headers.length - 1].getAttribute("spt_element_name");
-                spt.tab.select(last);
-            }
+            spt.tab.close(bvr.src_el); 
         '''
         } )
